test: split provider group removal coverage
Problem:
The grouped-removal regression coverage pushed an existing desktop test file beyond the repository size ceiling.

Root cause:
The new OpenCode Go tests were added to the already-large general provider-removal test file.

Fix:
Move the new grouped custom-provider tests into a focused test file without changing their behavior.

Verification:
- go run ./tools/repolint
- cd desktop && go test ./...
- focused grouped-removal tests
- focused grouped-removal tests under go test -race
